pyproject.toml
[project]
name = "video-player"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-vlc>=3.0.21203",
]
